// syncro/apps/web/src/lib/api/generated/model/machineRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Syncro API
 * Syncro backend REST API.
 * OpenAPI spec version: v1
 */
import type { MachineRequestStatus } from './machineRequestStatus';

export interface MachineRequest {
  plantId: string;
  machineGroupId: string;
  /**
     * @minLength 0
     * @maxLength 64
     * @pattern ^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$
     */
  code: string;
  /**
     * @minLength 0
     * @maxLength 255
     */
  name?: string;
  status: MachineRequestStatus;
  /**
     * @minLength 0
     * @maxLength 255
     */
  brand?: string;
  installedAt?: string;
  /**
     * @minLength 0
     * @maxLength 1000
     */
  notes?: string;
  /**
     * @minItems 0
     * @maxItems 10
     */
  optionalTelemetryFields?: string[];
}