// syncro/apps/web/src/lib/api/generated/model/machineView.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Syncro API
 * Syncro backend REST API.
 * OpenAPI spec version: v1
 */
import type { MachineViewStatus } from './machineViewStatus';
import type { TelemetryData } from './telemetryData';

export interface MachineView {
  id?: string;
  plantId?: string;
  plantCode?: string;
  plantName?: string;
  machineGroupId?: string;
  machineGroupName?: string;
  code?: string;
  name?: string;
  status?: MachineViewStatus;
  brand?: string;
  installedAt?: string;
  notes?: string;
  createdAt?: string;
  updatedAt?: string;
  optionalTelemetryFields?: string[];
  latestTelemetry?: TelemetryData;
}