// syncro/apps/web/src/lib/api/generated/model/machineViewStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Syncro API
 * Syncro backend REST API.
 * OpenAPI spec version: v1
 */

export type MachineViewStatus = typeof MachineViewStatus[keyof typeof MachineViewStatus];


export const MachineViewStatus = {
  ACTIVE: 'ACTIVE',
  INACTIVE: 'INACTIVE',
} as const;